put utils back in client.go
